creating the schemas and respective implementations

+++ src/main/java/com/iop/productservice/entity/Product.java
@@ -22,7 +22,7 @@ public class Product {
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
     @Column(name = "product_id")
-    private Integer productId;
+    private Long productId;
     @Column(name = "product_name")
     private String productName;
     @Column(name = "product_description")
