+++ src/main/java/com/iop/productservice/repository/ProductRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 
     // Named query
     Optional<Product> findByProductIdAndIsDeleted(Integer id, boolean isDeleted);
